refactor: 대기방 UI 개선 및 폴더 구조 정리 (#43)
- `GameIdCopyButton` 컴포넌트 생성 및 관련 함수 이동
- 폴더명 변경: `waiting-room` -> `waiting`
- 동적 라우팅 폴더명 변경: `[game-id]` -> `[gameId]`

코드 구조 개선 및 네이밍 일관성 향상
