+++ app/waiting/[gameId]/page.tsx
@@ -1,11 +1,10 @@
 'use client'
 
-import GameIdCopyButton from '@/components/features/waiting-room/game-id-copy-button'
+import GameIdCopyButton from '@/components/features/waiting/game-id-copy-button'
 import Button from '@/components/ui/button'
 import CatBox from '@/components/ui/cat-box'
-import useToastStore from '@/store/toast'
 
-const WaitingRoomPage = () => {
+const WaitingPage = () => {
   const GAME_ID = 'DJ31DK' // 예시 게임ID
 
   const players = [
@@ -17,14 +16,8 @@ const WaitingRoomPage = () => {
     { imageUrl: '', nickName: '' },
   ]
 
-  const { showToast } = useToastStore()
-
-  const handleCopyComplete = () => {
-    showToast('복사되었습니다', 'check')
-  }
-
   return (
-    <main className="md:bg-sea-spaceship-desk flex h-screen w-full flex-col items-center bg-sea-spaceship-mobile bg-cover bg-top">
+    <main className="flex h-screen w-full flex-col items-center bg-sea-spaceship-mobile bg-cover bg-top md:bg-sea-spaceship-desktop">
       <div className="mb-8 mt-[180px] grid grid-cols-3 grid-rows-2 gap-4">
         {players.map((player, index) => (
           <CatBox
@@ -35,7 +28,7 @@ const WaitingRoomPage = () => {
           />
         ))}
       </div>
-      <GameIdCopyButton gameId={GAME_ID} onCopySuccess={handleCopyComplete} />
+      <GameIdCopyButton gameId={GAME_ID} />
       <Button
         onClick={() => {
           console.log('게임 시작')
@@ -47,4 +40,4 @@ const WaitingRoomPage = () => {
   )
 }
 
-export default WaitingRoomPage
+export default WaitingPage
